Restore old behavior to disambiguate matching results
If two transformed variants have attributes such that for any given key they do not have a different value, we consider those
variants to be matching. We perform disambiguation by always choosing the last candidate, and then include any other
candidate which is not matching with the last candidate.

This in some way hides ambiguity, though we keep this behavior to not affect existing transform chains.